Allow builders to return default instance in Java
rationale: I was investigating allocations during parsing where a lot of empty messages and realized that each of these empty messages was a new instance, when it would be more efficient to reuse the default instance each time.

solution: add a default flag to the builder which is invalidated onChange. then conditionally return the default instance if that flag is true.